Bug fixes, tidy up test harness
Size of elements were not getting written into each object as it was read. Fix this by changed in SequenceHelper::Init to pass in a parameter.

Also, for implicit types, if they don't have any date, they aren't written, and have size = 0. This was throwing off the size calculation for TbsCertificate.

There is a bug in the Encode path where the sizes all get recalculated, and that leads to mischief. Need a better design to handle this case.